backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-community
langchain-core
langchain-groq
langchain-chroma
faiss-cpu
pypdf
wikipedia
arxiv
huggingface-hub
sentence-transformers
transformers
python-dotenv


